Add the hike story as an independent file
